# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Thailand Enterprise Accounting System API
  description: |
    REST API สำหรับระบบบัญชี enterprise ของไทย

    **Authentication:** API Key + HMAC signature, OR OAuth2 client_credentials
    **Idempotency:** ทุก POST/PUT ต้องมี Idempotency-Key header (UUID v4)
    **Rate limit:** 1,000 req/min per API key (default)
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
servers:
  # Host root only — paths carry their own prefix. BFF/JWT routes are bare
  # (e.g. /quotations); the external API key surface is under /api/v1/*; MCP at /mcp.
  - url: https://api.accounting.example.com
    description: Production
  - url: https://api-staging.accounting.example.com
    description: Staging
security:
  - ApiKeyAuth: []
  - OAuth2: [read, write]
paths:
...
  /admin/rbac/roles:
    get:
      summary: List roles for the target company (SUPER_ADMIN excluded) — perm sys.role.manage
      tags: [RBAC Admin]
      parameters:
        - { name: companyId, in: query, required: false, schema: { type: integer }, description: 'super-admin only; company-admin is forced to own company' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    roleId: { type: integer }
                    roleCode: { type: string }
                    nameTh: { type: string }
                    description: { type: string, nullable: true }
                    isSystem: { type: boolean }
                    userCount: { type: integer }
                    permissionCount: { type: integer }
        '403': { description: Forbidden (cross-company — rbac.cross_company.scope_required) }
    post:
      summary: Create a custom role in the target company — perm sys.role.manage
      tags: [RBAC Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleCode, nameTh]
              properties:
                roleCode: { type: string }
                nameTh: { type: string }
                description: { type: string, nullable: true }
                companyId: { type: integer, nullable: true, description: 'super-admin target; omitted = caller company' }
      responses:
        '201': { description: Created }
        '403': { description: Forbidden (cross-company) }
        '422': { description: Business rule (rbac.role_code_duplicate, rbac.super_admin_locked) }
  /admin/rbac/roles/{id}:
    get:
      summary: Role detail + permission codes — perm sys.role.manage
      tags: [RBAC Admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  roleId: { type: integer }
                  companyId: { type: integer }
                  roleCode: { type: string }
                  nameTh: { type: string }
                  description: { type: string, nullable: true }
                  isSystem: { type: boolean }
                  permissionCodes: { type: array, items: { type: string } }
        '404': { description: Not found (rbac.role.not_found) }
    put:
      summary: Rename a role (nameTh/description only) — perm sys.role.manage
      tags: [RBAC Admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nameTh]
              properties:
                nameTh: { type: string }
                description: { type: string, nullable: true }
      responses:
        '204': { description: Updated }
        '404': { description: Not found }
        '422': { description: SUPER_ADMIN locked (rbac.super_admin_locked) }
    delete:
      summary: Delete a role (not is_system, not in use) — perm sys.role.manage
      tags: [RBAC Admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }
        '422': { description: rbac.role_is_system / rbac.role_in_use / rbac.super_admin_locked }
  /admin/rbac/roles/{id}/permissions:
    put:
      summary: Replace a role's whole permission set (diff → audit rbac_grant_change) — perm sys.role.manage
      tags: [RBAC Admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissionCodes]
              properties:
                permissionCodes: { type: array, items: { type: string } }
      responses:
        '204': { description: Replaced }
        '404': { description: Not found }
        '422': { description: rbac.unknown_permission / rbac.super_admin_locked }
  /admin/rbac/users:
    get:
      summary: List users in the target company with their company-scoped roles — perm sys.user.manage
      tags: [RBAC Admin]
      parameters:
        - { name: companyId, in: query, required: false, schema: { type: integer } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    userId: { type: integer, format: int64 }
                    username: { type: string }
                    fullName: { type: string }
                    isActive: { type: boolean }
                    isSuperAdmin: { type: boolean }
                    roles:
                      type: array
                      items:
                        type: object
                        properties:
                          roleId: { type: integer }
                          roleCode: { type: string }
                          nameTh: { type: string }
        '403': { description: Forbidden (cross-company) }
    post:
      summary: Create a login user + assign initial company-scoped roles (never mints a super-admin) — perm sys.user.manage
      tags: [RBAC Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, fullName, isActive, roleIds]
              properties:
                username: { type: string, description: '3-64 chars [A-Za-z0-9_.-]' }
                password: { type: string, description: 'min 12 chars; never logged or returned' }
                fullName: { type: string }
                email: { type: string, nullable: true }
                isActive: { type: boolean }
                roleIds: { type: array, items: { type: integer } }
                companyId: { type: integer, nullable: true, description: 'super-admin target; omitted = caller company' }
      responses:
        '201': { description: 'Created (returns { userId })' }
        '403': { description: Forbidden (cross-company) }
        '422': { description: user.username_invalid / user.password_too_short / user.username_duplicate / rbac.role_company_mismatch }
  /admin/rbac/users/{id}/roles:
    put:
      summary: Replace a user's whole role set for the target company (audit user_role_change) — perm sys.user.manage
      tags: [RBAC Admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleIds]
              properties:
                roleIds: { type: array, items: { type: integer } }
                companyId: { type: integer, nullable: true, description: 'super-admin target; omitted = caller company' }
      responses:
        '204': { description: Replaced }
        '403': { description: Forbidden (cross-company) }
        '404': { description: User not found }
        '422': { description: rbac.role_company_mismatch / rbac.self_lockout }
  /admin/rbac/users/{id}/active:
    put:
      summary: Enable / disable a user account (deactivate = safe alternative to delete) — perm sys.user.manage
      tags: [RBAC Admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive: { type: boolean }
      responses:
        '204': { description: Updated }
        '403': { description: Forbidden (cross-company / super-admin) }
        '404': { description: User not found }
        '422': { description: rbac.self_lockout (cannot deactivate self) }
  /admin/rbac/users/{id}/password:
    put:
      summary: Admin-set a new password for a user (clears lockout/failed-count; never logged) — perm sys.user.manage
      tags: [RBAC Admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, description: 'min 12 chars; never logged or returned' }
      responses:
        '204': { description: Updated }
        '403': { description: Forbidden (cross-company / super-admin) }
        '404': { description: User not found }
        '422': { description: user.password_too_short }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Format: `key_<40-char-random>`. Issued via `POST /api-keys` (admin BFF).
        Plaintext returned once on create/rotate; only bcrypt hash + 16-char prefix stored.
        Scoped per-key (ScopesJson array of permission strings) + optional default_business_unit_id (auto-fill + lock per binding behavior).
        Required on all `/api/v1/*` endpoints (auth isolation: JWT not accepted on v1; ApiKey not accepted on root).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'BFF cookie-based JWT for human users via /auth/login. Required on all root endpoints (auth isolation: ApiKey not accepted on root).'
    OAuth2:
      type: oauth2
      description: 'Reserved for Phase 2 client_credentials flow when first customer demands OAuth. API key sufficient for Phase 1.'
      flows:
        clientCredentials:
          tokenUrl: https://api.your-teas.example/oauth/token
          scopes:
            read: Read access
            write: Write access
            admin: Admin access
...